Refactored Modal/Modal with flow
