--- src/components/Modal/Context.js
@@ -1,9 +1,0 @@
-// @flow
-import createContext from '@helpscout/react-utils/dist/createContext'
-import { noop } from '../../utilities/other'
-
-const Context = createContext({
-  closeModal: noop,
-})
-
-export default Context
