+++ src/components/Modal/__tests__/Footer.test.js
