+++ src/components/Portal/types.js
@@ -1,0 +1,14 @@
+// @flow
+
+export type PortalProps = {
+  className?: string,
+  exact?: boolean,
+  id?: string,
+  renderTo?: string | Object,
+  onBeforeOpen: () => void,
+  onOpen: () => void,
+  onBeforeClose: () => void,
+  onClose: () => void,
+  path?: string,
+  timeout: number,
+}
